requirements.txt
# ============================================================
# Pokemon Market Indexes - Python Dependencies
# ============================================================

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Supabase client (minimal)
postgrest>=0.16.0

# Date handling
python-dateutil>=2.8.0
